async_graphql_axum

Struct GraphQLWebSocket

pub struct GraphQLWebSocket<Sink, Stream, E, OnConnInit, OnPing> { /* private fields */ }

A Websocket connection for GraphQL subscription.

Implementations

impl<S, E> GraphQLWebSocket<SplitSink<S, Message>, SplitStream<S>, E, DefaultOnConnInitType, DefaultOnPingType>
where S: Stream<Item = Result<Message, Error>> + Sink<Message>, E: Executor,

pub fn new(stream: S, executor: E, protocol: GraphQLProtocol) -> Self

Create a GraphQLWebSocket object.

impl<Sink, Stream, E> GraphQLWebSocket<Sink, Stream, E, DefaultOnConnInitType, DefaultOnPingType>
where Sink: Sink<Message>, Stream: Stream<Item = Result<Message, Error>>, E: Executor,

pub fn new_with_pair( sink: Sink, stream: Stream, executor: E, protocol: GraphQLProtocol, ) -> Self

Create a GraphQLWebSocket object with sink and stream objects.

impl<Sink, Stream, E, OnConnInit, OnConnInitFut, OnPing, OnPingFut> GraphQLWebSocket<Sink, Stream, E, OnConnInit, OnPing>
where Sink: Sink<Message>, Stream: Stream<Item = Result<Message, Error>>, E: Executor, OnConnInit: FnOnce(Value) -> OnConnInitFut + Send + 'static, OnConnInitFut: Future<Output = Result<Data>> + Send + 'static, OnPing: FnOnce(Option<&Data>, Option<Value>) -> OnPingFut + Clone + Send + 'static, OnPingFut: Future<Output = Result<Option<Value>>> + Send + 'static,

pub fn with_data(self, data: Data) -> Self

Specify the initial subscription context data, usually you can get something from the incoming request to create it.

pub fn on_connection_init<F, R>( self, callback: F, ) -> GraphQLWebSocket<Sink, Stream, E, F, OnPing>
where F: FnOnce(Value) -> R + Send + 'static, R: Future<Output = Result<Data>> + Send + 'static,

Specify a callback function to be called when the connection is initialized.

You can get something from the payload of GQL_CONNECTION_INIT message to create Data. The data returned by this callback function will be merged with the data specified by [with_data].

pub fn on_ping<F, R>( self, callback: F, ) -> GraphQLWebSocket<Sink, Stream, E, OnConnInit, F>
where F: FnOnce(Option<&Data>, Option<Value>) -> R + Clone + Send + 'static, R: Future<Output = Result<Option<Value>>> + Send + 'static,

Specify a ping callback function.

This function if present, will be called with the data sent by the client in the Ping message.

The function should return the data to be sent in the Pong message.

NOTE: Only used for the graphql-ws protocol.

pub fn keepalive_timeout(self, timeout: impl Into<Option<Duration>>) -> Self

Sets a timeout for receiving an acknowledgement of the keep-alive ping.

If the ping is not acknowledged within the timeout, the connection will be closed.

NOTE: Only used for the graphql-ws protocol.

pub async fn serve(self)

Processing subscription requests.